More modular view resolution
